refactor: reoganized project files
