--- bistrain/tests/test_valid_a2c.yaml
@@ -1,44 +1,0 @@
-[GLOBAL]
-DEVICE = cpu
-SEED = 42
-AGENT = a2c
-ACTION_SIZE = 1
-ACTION_SPACE = continuous
-ACTION_RANGE = -2, 2
-STATE_SIZE = 3
-ENVIRONMENT = Pendulum-v0
-
-[A2C]
-
-  [[ACTOR]]
-  ARCHITECTURE = fc
-  HIDDEN_SIZE = 256, 32
-  LR = 0.001
-  WEIGHT_DECAY = 0.0
-  OPTIMIZER = adam
-
-  [[CRITIC]]
-  HIDDEN_SIZE = 256, 128
-  LR = 0.001
-  WEIGHT_DECAY = 1.0e-05
-  OPTIMIZER = adam
-
-  [[TRAINING]]
-  GAMMA = 0.99
-  N_STEP_BS = 4
-
-[EXPLORATION]
-TYPE = gaussian
-SIZE = 2
-EPS_BETA = 0.1
-EPS_MIN = 0.01
-MEAN = 0.0
-SIGMA = 0.4
-THETA = 0.01
-
-[TRAINER]
-MAX_STEPS = 200
-N_ENVS = 10
-EPISODES = 1000
-PRINT_EVERY =  100
-UPDATE_EVERY = 1
