+++ tests/test_valid_a2c.yaml
@@ -37,7 +37,7 @@ SIGMA = 0.4
 THETA = 0.01
 
 [TRAINER]
-MAX_T = 200
+MAX_STEPS = 200
 N_ENVS = 10
 EPISODES = 1000
 PRINT_EVERY =  100
